Pull CacheOptions out into its own service so that it can be used as a mixin along with other services
