+++ packages/sitecore-jss/src/i18n/dictionary-service.ts
@@ -1,35 +1,14 @@
-import { Cache, CacheClass } from 'memory-cache';
-
-const DEFAULTS = Object.freeze({
-  cacheTimeout: 60,
-  cacheEnabled: true,
-});
+import { CacheClient, CacheOptions, MemoryCacheClient } from '../cache-client';
 
 /**
- * Phrases from the Sitecore Dictionary Service
+ * Object model for Sitecore dictionary phrases
  */
 export interface DictionaryPhrases {
   [k: string]: string;
 }
 
 /**
- * Cache options (used by @see DictionaryService)
- */
-export interface CacheOptions {
-  /**
-   * Enable/disable caching mechanism
-   * @default true
-   */
-  cacheEnabled?: boolean;
-  /**
-   * Cache timeout (sec)
-   * @default 60
-   */
-  cacheTimeout?: number;
-}
-
-/**
- * Dictionary data fetching service
+ * Service that fetches dictionary data using Sitecore's GraphQL API.
  */
 export interface DictionaryService {
   /**
@@ -39,27 +18,49 @@ export interface DictionaryService {
   fetchDictionaryData(language: string): Promise<DictionaryPhrases>;
 }
 
-export abstract class DictionaryServiceBase implements DictionaryService {
-  private cache: CacheClass<string, DictionaryPhrases>;
+/**
+ * Base implementation of @see DictionaryService that handles caching dictionary values
+ */
+export abstract class DictionaryServiceBase
+  implements DictionaryService, CacheClient<DictionaryPhrases> {
+  private cache: CacheClient<DictionaryPhrases>;
 
+  /**
+   * Initializes a new instance of @see DictionaryService using the provided @see CacheOptions
+   * @param {CacheOptions} options Configuration options
+   */
   constructor(public options: CacheOptions) {
-    this.cache = new Cache();
-
-    this.options.cacheTimeout = (this.options.cacheTimeout ?? DEFAULTS.cacheTimeout) * 1000;
-
-    if (this.options.cacheEnabled === undefined) {
-      this.options.cacheEnabled = DEFAULTS.cacheEnabled;
-    }
+    this.cache = this.getCacheClient();
   }
 
-  protected getCacheValue(key: string): DictionaryPhrases | null {
-    return this.options.cacheEnabled ? this.cache.get(key) : null;
+  /**
+   * Caches a @see DictionaryPhrases value for the specified cache key.
+   * @param {string} key The cache key.
+   * @param {DictionaryPhrases} value The value to cache.
+   * @returns The value added to the cache.
+   * @mixes CacheClient<DictionaryPhrases>
+   */
+  setCacheValue(key: string, value: DictionaryPhrases): DictionaryPhrases {
+    return this.cache.setCacheValue(key, value);
   }
 
-  protected setCacheValue(key: string, value: DictionaryPhrases): DictionaryPhrases {
-    return this.options.cacheEnabled
-      ? this.cache.put(key, value, this.options.cacheTimeout)
-      : value;
+  /**
+   * Retrieves a @see DictionaryPhrases value from the cache.
+   * @param {string} key The cache key.
+   * @returns The @see DictionaryPhrases value, or null if the specified key is not found in the cache.
+   */
+  getCacheValue(key: string): DictionaryPhrases | null {
+    return this.cache.getCacheValue(key);
+  }
+
+  /**
+   * Gets a cache client that can cache data. Uses memory-cache as the default
+   * library for caching (@see MemoryCacheClient). Override this method if you
+   * want to use something else.
+   * @returns {CacheClient} implementation
+   */
+  protected getCacheClient(): CacheClient<DictionaryPhrases> {
+    return new MemoryCacheClient<DictionaryPhrases>(this.options);
   }
 
   /**
